pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-foundry"
version = "0.1.0"
description = "LLM Foundry Simulator - Stanford CS336 unified CLI pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "LLM Foundry Team"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "torch>=2.5",
    "numpy>=1.24",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "einops>=0.7.0",
    "jaxtyping>=0.2.25",
    "einx>=0.3.0",
    "regex>=2023.0",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "requests>=2.31",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["llm_foundry*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
